requirements.txt
Django==5.1.15
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
mysqlclient==2.2.8
gunicorn==23.0.0
